+++ portal/lib/app-version.ts
@@ -1,2 +1,2 @@
-export const APP_VERSION = "3.9.29";
-export const APP_BUILD_TIME = "2026-09-04";
+export const APP_VERSION = "3.9.30";
+export const APP_BUILD_TIME = "2026-09-05";